Send text to Messenger on boiler's status change
This functionality was added through send_text_to_messenger function.
To achieve that, some small changes were made to control_boiler function
and main.
The change to control_boiler was that it now returns a bool. That return value
is then used in main in order to decide whether an action to change the boiler's
status is actually needed or not.

Signed-off-by: Jason Pavlakis <jasonpavlakis@gmail.com>